Website formatting
Created a frame for the layout of the whole website. Added a section for the game description with an unrelated image as a test, a Deliverables section which will include links to the main deliverables stated in the assessment paper (e.g; pdfs of architecture), a section for Assets so that we can directly link to them on the website, and finally a larger section which would include any diagrams and charts we use, as they may not all be able to be placed in their respective deliverable without going over the page limit.